feat: move tests to the root
